updated lesson 5 and created lesson 6

+++ lesson5/main.c
@@ -1,27 +1,29 @@
 /**
  * Fichier principal du programme
- * Maintenant, le code se trouve à l'extérieur de main
+ * Maintenant, on appelle le code depuis main
+ * Code qui est désormais séparé en plusieurs fonctions distinctes,
+ * elles-mêmes situées à l'extérieure de la fonction principale.
  */
 
 #include <stdio.h>
 #include <stdlib.h>
 
-int add(int a, int b)
+float add(float a, float b)
 {
   return a + b;
 }
 
-int sub(int a, int b)
+float sub(float a, float b)
 {
   return a - b;
 }
 
-int mult(int a, int b)
+float mult(float a, float b)
 {
   return a * b;
 }
 
-int divide(double a, double b)
+float divide(float a, float b)
 {
   return a / b;
 }
@@ -29,7 +31,7 @@ int divide(double a, double b)
 int chooseCalc()
 {
   int chosedCalc = 0;
-  
+
   while (chosedCalc < 1 || chosedCalc > 5)
   {
     printf("\nWhat do you want to do :\n\n");
@@ -47,33 +49,33 @@ int chooseCalc()
 
 int main()
 {
-  int result = 0, numberOne = 0, numberTwo = 0;
+  float result = 0, numberOne = 0, numberTwo = 0;
   char chosed = '0';
   printf("Choose a number : ");
-  scanf("%d", &numberOne);
+  scanf("%f", &numberOne);
   printf("Another one : ");
-  scanf("%d", &numberTwo);
+  scanf("%f", &numberTwo);
   switch (chooseCalc())
   {
-    case 1:
-      result = add(numberOne, numberTwo);
-      chosed = '+';
-      break;
-    case 2:
-      result = sub(numberOne, numberTwo);
-      chosed = '-';
-      break;
-    case 3:
-      result = mult(numberOne, numberTwo);
-      chosed = '-';
-      break;
-    case 4:
-      result = divide(numberOne, numberTwo);
-      chosed = '/';
-      break;
-    case 5:
-      return 0;
+  case 1:
+    result = add(numberOne, numberTwo);
+    chosed = '+';
+    break;
+  case 2:
+    result = sub(numberOne, numberTwo);
+    chosed = '-';
+    break;
+  case 3:
+    result = mult(numberOne, numberTwo);
+    chosed = '-';
+    break;
+  case 4:
+    result = divide(numberOne, numberTwo);
+    chosed = '/';
+    break;
+  case 5:
+    return 0;
   }
-  printf("\nResult : %d %c %d = %d\n", numberOne, chosed, numberTwo, result);
+  printf("\nResult : %.0f %c %.0f = %.1f\n", numberOne, chosed, numberTwo, result);
   return 0;
 }
